store: extract relational tests to file
